Move bioimageio/ to micro_sam/v1/bioimageio submodule
